Isolating modules and separation of responsabilities
- The shared module should contains trully reusable components that
don't be singletons. They only import needed dependencies and should not
be used to populate imports of other modules, also shouldn't be imported
in AppModule or CoreModule

- The core module should contains standalone classes, models or services
that can be shared to the whole application, never components. Shouldn't
import SharedModule or any FeatureModule, and can be imported in any
other Modules.

- The block module should contains components that be singletons in the
application lifecycle or components that are created programmatically.
Can import CoreModule or SharedModule, but it never be imported in
SharedModule and it only should be imported int AppModule. An exception
are services that manage programmatically components, that import
directly the needed component file but never the whole module.

- The feature modules, only imports the required modules. They shouldn't export
any of their imports.