Set Auth Interceptor configurable
By default, the interceptor not inject any token. To get the token
injected, the AuthService config must declare regular expressions to set
witch request should be treated as protretected resource and the
Authorization token should be injected.

Also disabled the redirection to login by the interceptor, fixed a issue
getting the token when user isn't authenticated.