Make translation keys extractable and improve breadcrumb
Translation keys can be extracted automatically using
"extract-translation" npm task. The keys inside the code, to be detected by the
translation task, must be procesed by "mark" function and must be a
literal string.

LayoutService and Breadcrumb component are improved to separate title of
the page based on route. So now, the title of a route must be declared
in route data, using "title" attribute.

The translation file format are changed to be flat, and the "en" file
are removed because until now we don't support other language than "es".