Master detail forms layout
Initial implementation for sencond-level menu
--
Isolating functionality of each entity part

Each entity part is treated separatly, so the can be loaded and
persist independiently.

The component of each part must be linked to the correspondient
FormPart, and they is responsable of:

* Loading the neccesary data on first component load, future component
load (instantiation) uses the data from their assigned part.
* Retain in memory the data loaded
* Send the data to backend service, only that they handle!!

To coordinate all parts of an entity, a FormService must be extendend
and the needed parts are declared inside it.

As a requirement for all to work:

* The component that represents the edition or creation of an entity,
must extends MasterFormComponent. These component must provide the
FormService that each child routes (formParts) must use.

* Each child routes (formParts), must extends the DetailFormComponent
and itself request to the injected FormService for their managed
FormPart.

* For each DetailFormComponent must exists a FormPart that contains the
FormGroup and implements the loading of the necessary data.
 - For each MasterFormComponent must exists a FormService that have
declared all needed FormParts. For editions or creations the same
service can be used.
--
Separation between save and update. Fix status management

Now, save and update actions are managed separatly. So each part must
implement a separate logic for it.

The FormService treat this differently. For creations, all parts must
save without errors, when a error occurs the chain is aborted. During
the updating, the process continues so all parts are tried to be updated.

Also, status management are separated to better distinguishment between
creations and updates. The edition status is now published, and the
completed status is only meaningfully on creation
--
Add support to identity key propagation.
--
Cambios de ux
--
Fix translations
--
Adaptation of Evaluadores
--
Refactor to make more extensible and support for list element status

Renamed to fragment the part, and form service to ActionService. This
make it more generic. The Fragmet are splitted to allow fragments
without any form group, so that is the most basic fragmet.

Created a StatusWrapper to wrap the status of an object, usefull in
fragments where user modify a list of elements.
--
Adaptation of Evaluaciones
--
Adapt Gestion Evaluaciones
--
Adaptation of Convocatoria Reunion
--
Reorder menus
--
Remove cat module