Add management  of modules
The modules access is extracted from authorities, and published in
authStatus.

By convention, the first part is the module. For the special case of
authorities that end in '-INV' (ignoring the UO part), these are treated
as module INV and the prefix that identify the module is ignored.
So, authorities with suffix '-INV' are considered members of INV module.

Also added structural directives to check the access for a module and
published the functions for checking access (hasModuleAcces,
hasAnyModuleAccess).