Add last navigation between navigation in history state
The last navigation (known as from) can be accesed throught
history.state. This allow allow capturing it to returning the user to a
previous navigation.

The responsabilities of LayoutService are splitted, now
NavigationService manage all navigation states and the LayoutService use
it to propagate breadcrumb, title, and side menu changes.