Authenticated user authorities check inside auth service
Now the SgiAuthService provide methods to check the granted authorities
or modules of the user.

Are evaluated only when the user is authenticated, in other case always
returns false.

Changed directives and guards to use this methods instead of check
functions.

The base implementation now use these function to maintain isolation of
responsabilities and to allow simple override of the base
implementation.