Fix: ChromeHeadless have not captured in 60000 ms
The captureTimeout value represents the maximum boot-up time allowed
for a browser to start and connect to Karma. If any browser does not get
captured within the timeout, Karma will kill it and try to launch it again
and, after three attempts to capture it, Karma will give up.

Default: 60000
Updated to: 120000