Fix select validation
Sometimes select component wasn't reflect the validation status
(red highlight and validation error).

This have been fixed for all implementations, to guarantee a better
synchronization with underlaying reactive control.

For dependent select, when they have been touched by the user, the
status is forced to be coherent with the underlaying control validation.