Fix i18n input components and validators
On initializing the component, the existing logic generates multiple
changes in states updating multiple times value and generating redundant
callbacks that modify values quickly. This causes issues that was been masked in a Promise.

The logic have been reworked to guarantee an only one point of change
with their own responsability, chaining dependent callbacks only when is
absolutly neccessary.

Fixed the validation message that appeared before the control was touched
and never was hidden.

Fixed validators logic to guarantee consistency with Angular validators.

Renamed variables and methods to represent their objetives and clean
code.